Allow custom rich data types based on format
This makes it possible to utilize the `format` property to specify which
data type is to be used in the client or it should be translated
towards. Since this property is fully open ended, we are able to suit
any case we want.

The main change in this PR is starting to take into consideration this
value to discover which type the property gets translated to, before
only the actual type of the property was used (with binary being the
only `format` supported).

Only the v3 openApi generator was changed but the same changes can be
replicated to v2.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>