Generate personalised lune base client
This fully generates a lune base client whenever an operation is done.
The client has the name `luneClient.ts` and it currently mimics all
changes that are done to generate the final client. The final client is
identical to the currently generated one, but in this form it is
generated immediatelly at the generation step instead of having to check
which service/models files exist and build a file based on that.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>