Stop e2e testing on build
The e2e tests try to fully run an application as generated with a
client. This however, doesn't work for our custom case, since the
models/services generated require code present on the `lune-ts` library
itself. Since we already have unit tests for what the services/models
are expected to look, these tests value was already reduced, so they
were disabled. They also took a lot of time to produce so I guess builds
are now much faster.

In the future, we might want to reintroduce them. But this would either
require a more in close relationship with `lune-ts` by including it or
moving large chunks of it, or fully reworking/creating a new client type
that would correctly deal with it. Both these scenarios are quite work
heavy for now.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>