Fix unclosed comment on service template
We were not correclty closing the comment on the service itself.
Previously there was a `throws ApiError` that was rightly removed, but
the actual closing of the comment needs to remain.

This should have been caught with the changes in the templates
themselves, but there were so many changes introduced there that I
missed this one :see_no_evil:

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>