Add .js file extension to service imports
This is another of the custom Lune behaviours we want. This was already
previously done, but it was missing on the imports from the services
themselves. This is done to improve ESM/CommonJS compatibilities. It
also goes in line with the current state of `lune-ts` which we wanted to
fully replicate previously.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>