Fix luneClient template
The conversion to snake_case when sending the request is better suited
to be done when we create the body to send. This prevents having to
decode something already encoded in JSON just to make the conversion.
Additionally, the response conversion we want to fully maintain the same
format in both success and error, so simply fix that as well.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>