Add companion enum when using union types
This makes it that when we're using union types, we also add an enum
with the possible values in the models. This is necessary for Lune's
use case since union types cannot generate a list of values to
iterate through them, making things like get all values of this type
impossible. This is a use case we have in multiple places, for example
fetching all `CabinClass`. The import has also been added to the
`luneClient` whenever neededs: if the model is an enum and we're using
union types.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>