Make queryParams an object in services
The previous behaviour for the generation of methods in a service was to
use each parameter (query, path, body, cookie etc.) separately. This
meant that if we had 5 query parameters, the method would have 5
different arguments. Due to this, using methods with many of these
parameters can become cumbersome. Named parameters are not possible and
it brings issues where trying to provide a value for an optional
parameter require us to be explicit on previous optional parameters,
ending up with calls that are hard to understand and with requiring
explicit passing of the default values.

In an effort to have both better usage and allowing each method itself
to allow account overriding, this has been changed to treat queryParams
as a single object parameter. This allows for named parameters and makes
using the service method much clearer and concise.

For now, only query parameters have been moved to have this behaviour,
but one could easily do the same for any other type of parameter. For
our current schema though, the only other parameter type we use is path
and it's only a single parameter so this change feels unnecessary on
them.

The final goal is to have a service method something along the lines of:
```typescript
public method(bodyParams: XRequest, queryParams?: any, account: string)
```
Where `account` can be used to perform the operation sending a specific
account via the `Lune-account` header without modifying internal state,
preventing concurrency issues and facilitating usage for our customers
in the process.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>