Fix requestBody array handling
We need to correctly treat cases where the requestBody is not made of
properties but is a type itself. This happens whenever it references a
type that is an array, or if the content is straight up a type without
properties.

For these cases, handle it correctly and have it as an object if it is,
use it as a type if it isn't.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>