Rework methods interface (#26)
After some debate, we have decide to rework how we present the interface
for service methods. Two main changes are present here:
- Add option to override the account to be used on the endpoint for a
  single operation. This is made with a `options` object that is
optional and currently contains an `accountId` field. If we have more
to add in the future, the object structure allows us to easily do it.
- Move both queryParams and bodyParams to the same object. This makes it
  a lot easier for our clients to use and doesn't leak details about the
underlying API structure. This does require the parameters to not share
any of the property names, but this is already the case and we prefer
the tradeoff.

To note that the new `data` object that holds query and body parameters
is not nullable and it is optional if all parameters contained in it are
optional. The parameters inside `data` have also been sorted by required
status, mimicking the behaviour that existed for when they were outside.

An example of a possible interface with parameters in all places
(descriptions and comments are ommited):
```typescript
public methodX(
    pathParam1: string,
    pathParam2: string,
    data: {
        requestBodyParam1: typeX,
        queryParam2: number,
        requestBodyParam2?: typeZ,
        queryParam1?: string,
    },
    options?: { accountId: string },
)
```

Note: Some of the parser changes aren't reflected for a v2 OpenAPI
schema. This is because the v2 parser changes are a lot harder to do
and since we don't need it (our schema is >v3) we can simply ignore it.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>