Remove now obsolete method (#28)
The withAccount method was introduced to provide a way to perform an
operation having a different account in the `Lune-Account` header. Since
each method now has an `options` object that allows this override in a
thread safe way, this method is now obsolete. Remove it.

For cases where the purpose is to change the header in a more permanent
fashion, the `setAccount` method can be used.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>