Fix how default values interact with optional
The initial fix to the issue was correclty passing the default value of a
in the object parsing step itself, this was not correctly being
populated due to our custom made code.

However, this still didn't fully work since the normal handling of a
parameter assumes that when a default exists, the `?` modifier is not
needed since a default is to be provided. The result becomes:
`property: type = default`.

This can occur at parameters of the method, but since we use `properties`
as a way to have a single object parameter in the method (`data`), defaults
cannot exist inside it. The property was being set as not optional since
a default existed, but since the default cannot be populated it was
erratically just being marked as required.

To make sure that a default is still populated on parameters of the
method, but it is correctly only checking the `isRequired` flag for
properties inside `data`, create a `isRequiredStrict` and use it for
properties only (not parameters).

This fully fixes the case of `limit` as a query parameter that is not
required but does have a default. It is now an optional key inside
`data`.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>