Fix how default values interact with optional (#32)
The initial fix to the issue was correclty passing the default value of a
in the object parsing step itself, this was not correctly being
populated due to our custom made code.

However, this still didn't fully work since the normal handling of a
parameter assumes that when a default exists, the `?` modifier is not
needed since a default is to be provided. The result becomes:
`property: type = default`.

This can occur at parameters of the method, but since we use `properties`
as a way to have a single object parameter in the method (`data`), defaults
cannot exist inside it. The property was being set as not optional since
a default existed, but since the default cannot be populated it was
erratically just being marked as required.

To make sure that a default is still populated on parameters of the
method, but it is correctly only checking the `isRequired` flag for
properties inside `data`, create a `isRequiredStrict` and use it for
properties only (not parameters).

This fully fixes the case of `limit` as a query parameter that is not
required but does have a default. It is now an optional key inside
`data`.

Before/After (when default is present), showcasing it with the `limit`
parameter inside the `data` object:
**Before**
```typescript
methodX(data: { limit: string })
```
**After**
```typescript
methodX(data: { limit?: string })
```

A new template had to be created since if it's a direct parameter of the
method, we can populate with the default value and it is must be marked
as non optional. The behaviour with a default is:
```typescript
methodX(param1: typeZ = defaultValue)
```

If there was no need template and we simply reused `isRequired` these
parameters would also be affected and would become a wrong syntax:
```typescript
methodX(param1?: typeZ = defaultValue)
```

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>