Allow empty string as enum entry
This is a usecase we have in our OpenAPI schema. To allow it, the enum
is filtered out in the parsing stage, and the template deals with this
special case by naming the enum the empty string ('').

A new model was added to the tests that triggers the new code paths to
make sure generation is working.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>