Deal with null as an enum value (#37)
We have a hack in place where some of our enums have `null` as an enum
value. This is done to make our response validation middleware work
correctly. We need to deal with this new value accordingly. Since we
don't actually want to show it in the models, ignore any null as an enum
value. A test was added to verify the behaviour.

Keep in mind only one line was changed in the parsing logic, but github
seems to have not understood the change very efficiently :(

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>