Add uuid as a valid format (#43)
It seems we have introduced this format with the temporary_id from
payment, but the base generator was not correctly parsing it. This was
making it believe that 'string' was not a primitive type and was trying
to include a 'string.js' file. Simply add it to our existing list of
custom formats as a string.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>