Fix oneOf type at root level
Currently, if a oneOf model is present at the root level we try and use
its properties. However, these properties are usually unnammed making it
impossible to be used. Instead, we want to rely on the whole model as
the parameter itself. This is currently breaking our schema generation
since we've introduced a oneOf at the root.

We solve this by using the full model in the requestBody, making it use
the types of each oneOf. Additionally, a small fix to how this parameter
is sent to our internal clients also needed to be done. Tests were
added.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>