Bump version (patch) (#50)
* Fix allOf type at root level

We're having issues currently building lune-ts since we've added a allOf
at the root level. This is due to the way we parse properties and how we
deal with these properties in the service template. We already have code
in place that performs the necessary fixes, but it was only being done
for `oneOf`. Simply augment it to `allOf` at the root level as well.

New test was added.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>

* Bump version (patch)

This has the fixes for allOf at the root level.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>

---------

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>