Improve the type annotations in the generated client
error was explicitly of type "any". Similarly the "response" type was implicitly "any".

As a consequence of the former the compiler didn't notice quite a big
problem in the code[1] so we've been shipping bad code.

Let's provide type annotations and make them as narrow as possible to
reduce the chance of unnoticed issues left in the code.

Since we now know the type of response.data we need to silence the
compiler errors on that one particular line – the comment explains why.

[1] https://github.com/lune-climate/openapi-typescript-codegen/pull/51