Update test snapshots
The current snapshot that is expected in the tests is not up to date,
meaning the tests fail. This PR updates the snapshot to be in sync with
the current version of the code, no functional changes were made.

This was making subsequent changes to the actual logic hard to read
since updating the tests brings also these updates that were missing.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>