Fix case where query parameters are not added (#61)
We've introduced query parameters in the generator by adding them as
normal parameters in the request body for our users. This was done by
simply pushing these parameters to the `data` field the users use to
send the necessary data to the endpoint.

However, there was an edge case. Since we first process the query params
and then the requestBody, for cases where the requestBody was a
`anyOf`/`allOf` we were fully substituting the parameters to be the
requestBody, overwriting the query parameters that were previously
added. To fix this, simply inverse the order of operations, first add
the requestBody to the parameters and then process query parameters.

This made some existing tests change the order of the parameters inside
the object but resulting in the same data in the end. A test case was added
for this particular scenario.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>