Fix `data` being wrongly considered optional
Due to the latest change to fix how query parameters are handled, the
`isRequired` flag on the internal model was not being correctly
preserved. If the requestBody is mandatory then the whole `data` object
is mandatory, same for query parameters. We were wrongly overwriting
`isRequired` to whatever the query params stated, but we want to
preserve this information since the requestBody has been processed at
this point now.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>