Rework logic
We have decided to not have per-endpoint versions nor client wide
versions. Instead, we will simply have a single version for all the
types that needs to be specified in the request as a header. Rename the
parameter to instead receive a single API Version which is stored in the
`luneClient` configuration, which can be used by `lune-ts` to send via
a header.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>