Add API version value to configuration (#73)
Add a new required parameter to the application that specifies the
API calendar versions that is to be used. This value is currently added
to the configuration of the generated client `luneClient.ts` so it can be
used later. The goal is to have all requests in `lune-ts` add this value in
the `Lune-Version` header so the library works on a single version,
always matching the code with the types in place.

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>