Support `https-url` as format (#77)
We have a custom made mapping between OpenAPI formats and the
corresponding types. We have started using the `https-url` format in the
API schema but didn't yet support it internally which was making it be
treated as not a primitive type resulting in an import of the `string`
type (which fails [^1]).

Add support for this format and treat it as a normal string in terms of
type. Tests were added/adjusted accordingly.

[^1]: https://github.com/lune-climate/lune-ts/actions/runs/11670865015/job/32495994689

Signed-off-by: Ruben Aguiar <r.aguiar9080@gmail.com>